fix: 모든 input 데이터 유효성 검사 및 DTO 통일성 유지하기 완료

+++ src/api/wishlist/wishlist.service.ts
@@ -1,27 +1,18 @@
 import { Injectable } from '@nestjs/common';
 import { WishlistRepository } from './wishlist.repository';
-import { ShopIdDto } from './dto/wishlist.dto';
+import { SaveWishListDTO } from './dto/wishlist.dto';
 
 @Injectable()
 export class WishlistService {
   constructor(private wishlistRepository: WishlistRepository) {}
 
-  async addWishlistByShopIdAndUUID(shopIdDto: ShopIdDto, uuid: string) {
-    const { shopId } = shopIdDto;
+  async addWishlistByShopIdAndUUID(saveWishListDTO: SaveWishListDTO, uuid: string) {
+    const { shopId } = saveWishListDTO;
 
-    const wishlist = await this.wishlistRepository.findWishlistByShopIdAndUUID(
-      shopId,
-      uuid,
-    );
-    if (wishlist) {
-      return await this.wishlistRepository.deleteWishlistByWishlistId(
-        wishlist.id,
-      );
-    }
-    const result = await this.wishlistRepository.addWishlistByShopIdAndUUID(
-      shopId,
-      uuid,
-    );
+    const wishlist = await this.wishlistRepository.findWishlistByShopIdAndUUID(shopId, uuid);
+    if (wishlist) return await this.wishlistRepository.deleteWishlistByWishlistId(wishlist.id);
+
+    const result = await this.wishlistRepository.addWishlistByShopIdAndUUID(shopId, uuid);
     return result;
   }
 }
